refactor CommonKeys (#1825)
