+++ tests/test_threadcontainer.py
@@ -15,11 +15,12 @@ import unittest
 import torch
 
 from monai.utils import optional_import
+from monai.utils.enums import CommonKeys
 
 try:
     _, has_ignite = optional_import("ignite")
 
-    from monai.engines import CommonKeys, SupervisedTrainer
+    from monai.engines import SupervisedTrainer
     from monai.utils import ThreadContainer
 except ImportError:
     has_ignite = False
